conf: Set source type of the stub console
from Osier Yang <jyang@redhat.com>
------------------------------------------------------------------------

As "serials[0]->source.type". This is for kvm tool driver only
supports one console currently, and it setup ptys (redirect
the stdout/stderr of kvm tool process to the master pty, and
the slave pty path will be set to source path of the console)
for connect to the guest via console. If the console config
is like:

      <console type='pty'>
        <target type='serial' port='0'/>
      </console>

Things don't work, as kvm tool driver only tries to find consoles
and set the slave pty path to "consoles[0]->source.path". Though
another solution is to set the slave pty path to "serials[0]->source.path"
if no console of type "pty" can be found.